requirements.txt
pyrogram
tgcrypto
aiohttp
mutagen
python-dotenv
Pillow
psycopg2-binary
pathvalidate
pycryptodomex
psutil